Cleaned up the code to handle changes in Oxalis

+++ ringo-persistence-jdbc-impl/src/main/java/no/sr/ringo/document/DocumentRepositoryImpl.java
@@ -54,7 +54,7 @@ public class DocumentRepositoryImpl implements DocumentRepository {
         if (documentFound(rs)) {
             return extractPeppolDocumentFromResultSet(rs);
         } else {
-            throw new PeppolMessageNotFoundException(msgNo.toLong());
+            throw new PeppolMessageNotFoundException(msgNo);
         }
     }
